app/requirements.txt
flask
psycopg2-binary
requests
docling
